% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_mat.R
\name{distance_mat}
\alias{distance_mat}
\title{Distance Matrix Function}
\usage{
distance_mat(diags)
}
\arguments{
\item{diags}{A list of persistence diagrams.}
}
\description{
This function computes a distance matrix between all persistence diagrams using the Wasserstein distance.
}
\examples{
distance_mat()
}
