% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointloss.R
\name{jointloss}
\alias{jointloss}
\title{Joint Loss Function}
\usage{
jointloss(distancemat, grouplabels)
}
\arguments{
\item{distancemat}{A distance matrix, the output of distance_mat().}

\item{grouplabels}{A vector of labels for one of the two groups.}
}
\description{
This function computes our test statistic under the given labelling scheme.
}
\examples{
jointloss()
}
