% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_sim_block.R
\name{montecarlo_sim_block}
\alias{montecarlo_sim_block}
\title{Monte Carlo Simulation Function}
\usage{
montecarlo_sim_block(distancemat, grouplabels, N, blocks, blocksize)
}
\arguments{
\item{distancemat}{A distance matrix. The output of distance_mat()}

\item{grouplabels}{A set of labels(indices) for one of the two groups in experiment.}
  
\item{levels}{Two-column matrix encoding index information for two-level point-cloud grouping. First column is whole-block shuffling partition and second column is within-block shuffling partition.}

\item{N}{The number of permutations to be performed.}

}
\description{
This function computes a p-value using a monte carlo simulation with block sampling.
}
\examples{
montecarlo_sim_block()
}
