% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/montecarlo_sim_block.R
\name{montecarlo_sim_block}
\alias{montecarlo_sim_block}
\title{Monte Carlo Simulation Function}
\usage{
montecarlo_sim_block(distancemat, grouplabels, N, blocks, blocksize)
}
\arguments{
\item{distancemat}{A distance matrix. The output of distance_mat()}

\item{grouplabels}{A set of labels(indices) for one of the two groups in experiment.}

\item{N}{The number of permutations to be performed.}

\item{blocks}{The number of exchangeable blocks. Usually the number of epochs in an experiment.}

\item{blocksize}{The size of each block, only give one value.}
}
\description{
This function computes a p-value using a monte carlo simulation with block sampling.
}
\examples{
montecarlo_sim_block()
}
